Fix home page and overflow
