+++ study_app_client/.dart_tool/version
@@ -1,1 +1,1 @@
-3.32.5
+3.32.8
